% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_gq.R
\name{age_dens}
\alias{age_dens}
\alias{ages_edens}
\alias{lambda_drm}
\title{Age-specific expected densities based on DRM.}
\usage{
ages_edens(drm, cores = 1)

lambda_drm(...)
}
\arguments{
\item{drm}{A \code{list} object containing the output from the \link{fit_drm}
function.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{drm}, then four (or less) threads are recommended.}

\item{...}{params to be passed to \code{age_edens}}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for forecasting.
}
\description{
Age-specific expected densities based on DRM.
}
\author{
lcgodoy
}
