% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{between_scalar}
\alias{between_scalar}
\title{Check if elements in x are between corresponding elements in lb and
ub}
\usage{
between_scalar(x, lb, ub)
}
\arguments{
\item{x}{A numeric vector.}

\item{lb}{A numeric vector of lower bounds.}

\item{ub}{A numeric vector of upper bounds.}
}
\value{
A logical vector of the same length as x, indicating whether each
element of x is between the corresponding elements of lb and ub.
}
\description{
Check if elements in x are between corresponding elements in lb and
ub
}
\examples{
between(1:5, 1, 5)
between(1:5, 2, 4)

}
