% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_between}
\alias{check_between}
\title{Check if x is between lb and ub}
\usage{
check_between(x, lb, ub)
}
\arguments{
\item{x}{A numeric vector.}

\item{lb}{A numeric vector of lower bounds.}

\item{ub}{A numeric vector of upper bounds.}
}
\value{
Returns \code{NULL} invisibly. This function stops execution if any of the
following conditions are met:
\itemize{
\item lb is greater than or equal to ub.
\item The lengths of lb and ub are not equal.
\item The lengths of lb, ub, and x are not equal.
}
}
\description{
Check if x is between lb and ub
}
