% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{check_pars}
\alias{check_pars}
\title{Validate input data for \code{\link[=prior_sample]{prior_sample()}}}
\usage{
check_pars(dat, model)
}
\arguments{
\item{dat}{A \code{list} containing the data, typically generated by
\code{\link[=make_data]{make_data()}} or \code{\link[=make_data_sdm]{make_data_sdm()}}.}

\item{model}{A \code{character} string specifying the model type. Must be
either "drm" or "sdm". Defaults to "drm".}
}
\value{
This function does not return a value. It stops execution with an
error message if the data is not valid for the specified model.
}
\description{
Checks if the provided data list is valid for a given model.
This is an internal function not intended for direct use.
}
\seealso{
\code{\link[=make_data]{make_data()}}, \code{\link[=make_data_sdm]{make_data_sdm()}}
}
\author{
lcgodoy
}
\keyword{internal}
