% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{clean_edens}
\alias{clean_edens}
\title{Cleaning the variable output from stan for interpretability.}
\usage{
clean_edens(x)
}
\arguments{
\item{x}{A string}
}
\description{
Cleaning the variable output from stan for interpretability.
}
