% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_laplace}
\alias{default_laplace}
\title{Default Laplace arguments}
\usage{
default_laplace()
}
\value{
A \code{list} containing the default arguments for the Laplace algorithm.
}
\description{
Returns default for Laplace
}
\author{
lcgodoy
}
