% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_nuts}
\alias{default_nuts}
\title{Default NUTS arguments}
\usage{
default_nuts()
}
\value{
A \code{list} containing the default arguments for the NUTS algorithm.
}
\description{
Returns default for NUTS
}
\author{
lcgodoy
}
