% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_opt}
\alias{default_opt}
\title{Default Optimization arguments}
\usage{
default_opt()
}
\value{
A \code{list} containing the default arguments for the Optimization algorithm.
}
\description{
Returns default for Optimization
}
\author{
lcgodoy
}
