% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_priors}
\alias{default_priors}
\title{Default priors' hyperparameters}
\usage{
default_priors()
}
\value{
A \code{list} containing the default hyperparameter values for the priors.
}
\description{
Returns default priors' hyperparameters for the DRM model.
}
\author{
lcgodoy
}
