% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{default_qt}
\alias{default_qt}
\title{Default quantities}
\usage{
default_qt()
}
\value{
a `list` with (explain values)
}
\description{
Returns default quantities for the DRM model.
}
\details{
Based on Alexa's Summer Flounder project.
}
\author{
lcgodoy
}
