% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_vb}
\alias{default_vb}
\title{Default VB arguments}
\usage{
default_vb()
}
\value{
A \code{list} containing the default arguments for the Variational Bayes algorithm.
}
\description{
Returns default for Variational Bayes
}
\author{
lcgodoy
}
