% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.get_param_description}
\alias{.get_param_description}
\title{Internal helper to generate parameter descriptions}
\usage{
.get_param_description(v, covars, model = c("drm", "sdm"))
}
\arguments{
\item{v}{A character string representing the parameter name}

\item{covars}{A list of covariate names extracted from the model object}

\item{model}{A list of covariate names extracted from the model object}
}
\description{
Internal helper to generate parameter descriptions
}
