% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{draws}
\alias{draws}
\alias{draws.drmrmodels}
\title{Draws method for \code{adrm} and \code{sdm} objects.}
\usage{
draws(x, ...)

\method{draws}{drmrmodels}(
  x,
  variables = NULL,
  inc_warmup = FALSE,
  format = getOption("cmdstanr_draws_format", "draws_array"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{adrm} or \code{sdm}.}

\item{...}{currently ignored.}

\item{variables}{a string vector with the name of the variables for which we
want to obtain posteriors samples (defaults to the same variables as the
\code{summary} methods).}

\item{inc_warmup}{a boolean indicating whether the warmup samples should be
retrieved as well. Defaults to \code{FALSE}.}

\item{format}{A string. See cmdstanr \href{https://mc-stan.org/cmdstanr/reference/fit-method-draws.html}{$draws documentation}.}
}
\value{
an object of class \code{"draws"} containing the posterior samples
from specified parameters.
}
\description{
Mirrors the behavior of \code{cmdstanr}'s \code{$draws} method
to retrieve the samples from the posterior distribution of the models'
parameters.
}
\author{
lcgodoy
}
