% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dtn}
\alias{dtn}
\title{Density of a truncated Normal distribution}
\usage{
dtn(x, mean = 0, sd = 1, range = c(0, Inf), log = FALSE)
}
\arguments{
\item{x}{A \code{numeric vector} at which to evaluate the density.}

\item{mean}{A \code{numeric} scalar representing the mean of the underlying
normal distribution. Defaults to 0.}

\item{sd}{A \code{numeric} scalar representing the standard deviation of the
underlying normal distribution. Defaults to 1.}

\item{range}{A \code{numeric} vector of length 2 specifying the lower and
upper truncation bounds. Defaults to \code{c(0, Inf)}, indicating truncation
from below at 0.}

\item{log}{\code{Logical}; if \code{TRUE}, the natural logarithm of the
density is returned. Defaults to \code{FALSE}.}
}
\value{
A numeric vector of the same length as \code{x}, containing the
(log) density values of the truncated normal distribution.
}
\description{
Calculates the density of a truncated Normal distribution.
}
\seealso{
\code{\link[stats:dnorm]{stats::dnorm()}}, \code{\link[stats:pnorm]{stats::pnorm()}}
}
\author{
lcgodoy
}
