% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd_gq.R
\name{elpd}
\alias{elpd}
\alias{elpd.adrm}
\alias{elpd.sdm}
\title{(out-of-sample) Expected Log-posterior Density (ELPD) based on
\code{adrm} and \code{sdm} objects.}
\usage{
elpd(x, ...)

\method{elpd}{adrm}(x, new_data, past_data, f_test, seed = 1, cores = 1, ...)

\method{elpd}{sdm}(x, new_data, seed = 1, cores = 1, ...)
}
\arguments{
\item{x}{A \code{list} object containing the output from the \code{\link[=fit_drm]{fit_drm()}} (or
\code{\link[=fit_sdm]{fit_sdm()}}) function.}

\item{...}{parameters to be passed to \code{elpd}}

\item{new_data}{a \code{data.frame} with the dataset at which we wish to
obtain predictions. Note that, this \code{data.frame} must contain the
response variable used when fitting the DRM as well.}

\item{past_data}{a \code{data.frame} with the dataset last year used in
model fitting. Only needed when \code{f_test} is not missing or when
estimating survival.}

\item{f_test}{a \code{matrix} informing the instantaneous fishing mortality
rates at each age (columns) and timepoint (rows).}

\item{seed}{a seed used for the forecasts. Forecasts are obtained through
Monte Carlo samples from the posterior predictive distribution. Therefore,
a \code{seed} is needed to ensure the results' reproducibility.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{drm}, then four (or less) threads are recommended.}
}
\value{
an object of class \code{"CmdStanGQ"} containing the ELPD
function evaluated at each data point given each sample from the
posterior.
}
\description{
Considering a new dataset (across the same patches), computes
the out-of-sample ELPD based on the DRM passed as \code{drm}.
}
\details{
The current version of the code assumes the data where forecasts
are needed are ordered by "patch" and "site" and, in addition, the patches
MUST be the same as the ones used to obtain the parameters' estimates from
the the \code{drm} object.
}
\author{
lcgodoy
}
