% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd_gq.R
\name{elpd_sdm}
\alias{elpd_sdm}
\title{ELPD based on SDM.}
\usage{
elpd_sdm(sdm, new_data, seed = 1, cores = 1)
}
\arguments{
\item{sdm}{A \code{list} object containing the output of a \link{fit_sdm} call.}

\item{new_data}{a \code{data.frame} with the dataset at which we wish to
obtain predictions.}

\item{seed}{a seed used for the forecasts. Forecasts are obtained through
Monte Carlo samples from the posterior predictive distribution. Therefore,
a \code{seed} is needed to ensure the results' reproducibility.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{drm}, then four (or less) threads are recommended.}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for forecasting.
}
\description{
Considering a new dataset (across the same patches), computes
ELPD based on the SDM passed as \code{sdm}. Note that, the response
variable must be available in the \code{new_data} as well.

Consider a linear predictor having linear and square terms
associated with a variable \eqn{x}. Assume this variable was centered
before being included in the linear predictor. This functions returns the
value of \eqn{x} (on its original scale) such that the linear predictor is
maximized (or minimized).
}
\details{
The current version of the code assumes the data where forecasts
are needed is ordered by "patch" and "site" and, in addition, its patches
MUST be the same as the ones used to obtain the parameters' estimates from
the the \code{sdm} object.
}
\author{
lcgodoy
}
