% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_gq.R
\name{fitted.adrm}
\alias{fitted.adrm}
\title{DRM fitted values}
\usage{
\method{fitted}{adrm}(object, cores = 1, ...)
}
\arguments{
\item{object}{\code{adrm} object containing the output from the \link{fit_drm}
function.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{fit_drm}, then four (or less) threads are recommended.}

\item{...}{additional parameters to be passed to \code{$generated_quantities}}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for the observed data.
}
\description{
Calculates the "fitted values" associated to an \code{adrm}
object. These fitted values are useful for "hindcasts" and posterior
predictive checks.
}
\author{
lcgodoy
}
