% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_gq.R
\name{fitted_pars_ll}
\alias{fitted_pars_ll}
\title{Retrieve parameters needed for forecasting}
\usage{
fitted_pars_ll(data_list)
}
\arguments{
\item{data_list}{a \code{list} used as input for model fitting. Typically,
the output from the \link{make_data} function.}
}
\value{
a \code{character} vector of labels indicating the parameters
necessary for the forecast.
}
\description{
This function identifies the parameters necessary for carrying
out forecasts based on the data used to fit a DRM (or SDM).
}
\author{
lcgodoy
}
