% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_dd}
\alias{fix_dd}
\title{Density dependence verbose to code}
\usage{
fix_dd(x)
}
\arguments{
\item{x}{a \code{character}}
}
\value{
An \code{integer}.
}
\description{
Density dependence verbose to code
}
\author{
lcgodoy
}
