% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_linbeta}
\alias{fix_linbeta}
\title{Regression coefficient for non-centered variable}
\usage{
fix_linbeta(beta1, beta2, offset)
}
\arguments{
\item{beta1}{A \code{numeric} regression coefficient associated to the
linear term.}

\item{beta2}{A \code{numeric} regression coefficient associated to the
quadratic term.}

\item{offset}{a \code{numeric} representing the "center" of \eqn{x}.}
}
\value{
a \code{numeric} representing the regression coefficient of the
linear term for the model where \eqn{x} is not centered.
}
\description{
Consider a linear predictor having linear and square terms
associated with a variable \eqn{x}. Assume this variable was centered
before being included in the linear predictor. This functions recovers the
regression coefficient associated to the linear term as if the variable
was not centered.
}
\author{
Lucas Godoy
}
