% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_re}
\alias{fix_re}
\title{Random effects verbose to code}
\usage{
fix_re(x)
}
\arguments{
\item{x}{a \code{character}}
}
\value{
a \code{integer}
}
\description{
Random effects verbose to code
}
\author{
lcgodoy
}
