% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{gen_adj}
\alias{gen_adj}
\title{Generates an adjacency matrix}
\usage{
gen_adj(x)
}
\arguments{
\item{x}{an \code{sf} object representing the patches.}
}
\value{
An adjacency \code{matrix}.
}
\description{
Generates an adjacency matrix
}
\author{
lcgodoy
}
