% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_scaling}
\alias{get_scaling}
\title{Scaling factor for ICAR}
\usage{
get_scaling(adj)
}
\arguments{
\item{adj}{adjacency \code{matrix}}
}
\value{
a \code{scalar}
}
\description{
Using results from Rue and Held 2005 and Morris et al. 2019.
}
\author{
lcgodoy
}
