% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_zeros}
\alias{get_zeros}
\title{Zeros and nonzeros}
\usage{
get_zeros(y)
}
\arguments{
\item{y}{a \code{numeric vector} of species' densities.}
}
\value{
A \code{list} containing necessary input for stan to identify zeros
and nonzeros in the response variable.
}
\description{
Zeros and nonzeros
}
\author{
lcgodoy
}
