% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ginv}
\alias{ginv}
\title{Generalized Inverse}
\usage{
ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{A matrix we wish to invert.}

\item{tol}{A relative tolerance to detect zero singular values.}
}
\value{
The generalized inverse of \code{X}
}
\description{
Generalized Inverse
}
\details{
this function is taken from the package \code{MASS}
}
