% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{lambda2df}
\alias{lambda2df}
\title{Turn an array of density per age, time, and patch/site into a
\code{data.frame}}
\usage{
lambda2df(lbd)
}
\arguments{
\item{lbd}{a 3-dimensional \code{array}}
}
\value{
return a \code{data.frame}.
}
\description{
Turn an array of density per age, time, and patch/site into a
\code{data.frame}
}
\author{
lcgodoy
}
