% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_gq.R
\name{lambda_drm}
\alias{lambda_drm}
\title{Age-specific densities based on DRM.}
\usage{
lambda_drm(drm, cores = 1)
}
\arguments{
\item{drm}{A \code{list} object containing the output from the \link{fit_drm}
function.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{drm}, then four (or less) threads are recommended.}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for forecasting.
}
\description{
Considering a new dataset (across the same patches), computes
forecasts based on the DRM passed as \code{drm}.
}
\author{
anonymous
}
