% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_gq.R
\name{loglik}
\alias{loglik}
\alias{log_lik}
\alias{log_lik.adrm}
\alias{log_lik.sdm}
\title{Computing the log-likelihood function fot \code{adrm} and \code{sdm} objets.}
\usage{
log_lik(x, cores = 1)

\method{log_lik}{adrm}(x, cores = 1)

\method{log_lik}{sdm}(x, cores = 1)
}
\arguments{
\item{x}{an object of class \code{adrm} or \code{sdm}. These objects are
generated as the output of the \code{\link[=fit_drm]{fit_drm()}} and \code{\link[=fit_sdm]{fit_sdm()}} functions,
respectively,}

\item{cores}{number of threads used to calculate the log-likelihood
function. If four chains were used in the \code{fit_drm} (of
\code{fit_sdm}), then four (or less) threads are recommended.}
}
\value{
an object of class \code{"CmdStanGQ"} containing the log-likelihood
function evaluated at each data point given each sample from the
posterior.
}
\description{
Computing the log-likelihood function fot \code{adrm} and \code{sdm} objets.
}
\author{
lcgodoy
}
