% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_gq.R
\name{loglik_drm}
\alias{loglik_drm}
\title{DRM log-likelihood (for model comparisons' purpose)}
\usage{
loglik_drm(drm, cores = 1)
}
\arguments{
\item{drm}{A \code{list} object containing the output from the \link{fit_drm}
function.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{drm}, then four (or less) threads are recommended.}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for forecasting.
}
\description{
Calculates the log-likelihood associated to a given \code{drm}
model.
}
\author{
lcgodoy
}
