% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{make_surv}
\alias{make_surv}
\title{Generate the "survival" terms}
\usage{
make_surv(n_patches, n_time, x_sv, pars)
}
\arguments{
\item{n_patches}{number of patches}

\item{n_time}{number of timepoints}

\item{x_sv}{matrix of environmental factors affecting survival}

\item{pars}{a named \code{list} of parameters used to simulate
log-recruitment. It must contain a vector named \code{"beta_s"} with
length equal to the number of columnts in \code{x_sv}.}
}
\value{
a \code{matrix} with \code{n_patches} columns and \code{n_time} rows
representing the log-survival at each patch/site and time.
}
\description{
Generate the "survival" terms
}
\author{
lcgodoy
}
