% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_quad_x}
\alias{max_quad_x}
\title{Value of a covariate that maximizes the response variable in a
quadratic model.}
\usage{
max_quad_x(beta1, beta2, offset = 0)
}
\arguments{
\item{beta1}{A \code{numeric} regression coefficient associated to the
linear term.}

\item{beta2}{A \code{numeric} regression coefficient associated to the
quadratic term.}

\item{offset}{a \code{numeric} representing the "center" of \eqn{x}.}
}
\value{
a \code{numeric} representing the uncentered \eqn{x} that maximizes
(or minimizes) the linear predictor.
}
\description{
Consider a linear predictor having linear and square terms
associated with a variable \eqn{x}. Assume this variable was centered
before being included in the linear predictor. This functions returns the
value of \eqn{x} (on its original scale) such that the linear predictor is
maximized (or minimized).
}
\author{
anonymous
}
