% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mcmc_diag}
\alias{mcmc_diag}
\alias{mcmc_diag.drmrmodels}
\title{Draws method for \code{adrm} and \code{sdm} objects.}
\usage{
mcmc_diag(x, ...)

\method{mcmc_diag}{drmrmodels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{adrm} or \code{sdm}.}

\item{...}{currently ignored.}
}
\value{
an object of class \code{"drmrdiag"}
}
\description{
Mirrors the behavior of \code{cmdstanr}'s \code{$draws} method
to retrieve the samples from the posterior distribution of the models'
parameters.
}
\author{
lcgodoy
}
