% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_adrm}
\alias{new_adrm}
\title{Create an adrm object}
\usage{
new_adrm(x = list())
}
\arguments{
\item{x}{A \code{list} that has the required fields for an 'adrm' object.
Defaults to an empty \code{list}.}
}
\value{
An object of class \code{adrm}.
}
\description{
Constructs an object of class \code{adrm} from a list, after validating its
structure. Note that, this function is mostly for internal usage. \code{adrm}
stands for age-structured dynamic range model.
}
