% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_aesd}
\alias{new_aesd}
\title{Create a aesd object}
\usage{
new_aesd(x = list())
}
\arguments{
\item{x}{A \code{list} that has the required fields for a 'aesd' object.
Defaults to an empty \code{list}.}
}
\value{
An object of class \code{aesd}.
}
\description{
Constructs an object of class \code{aesd} from a list, after validating its
structure. Note that this function is mostly for internal usage.
}
