% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_pred_drmr}
\alias{new_pred_drmr}
\title{Create a pred_drmr object}
\usage{
new_pred_drmr(x = list())
}
\arguments{
\item{x}{A \code{list} that has the required fields for a 'pred_drmr' object.
Defaults to an empty \code{list}.}
}
\value{
An object of class \code{pred_drmr}.
}
\description{
Constructs an object of class \code{pred_drmr} from a list, after validating its
structure. Note that this function is mostly for internal usage.
}
