% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_sdm}
\alias{new_sdm}
\title{Create an sdm object}
\usage{
new_sdm(x = list())
}
\arguments{
\item{x}{A \code{list} that has the required fields for an 'sdm' object.
Defaults to an empty \code{list}.}
}
\value{
An object of class \code{sdm}.
}
\description{
Constructs an object of class \code{sdm} from a list, after validating its
structure. Note that, this function is mostly for internal usage. \code{sdm}
stands for species distribution model.
}
