% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{pars_transform}
\alias{pars_transform}
\title{Transform parameters to a meaningful and interpretable scale.}
\usage{
pars_transform(pars)
}
\arguments{
\item{pars}{\code{list} of parameters}
}
\value{
\code{list} of parameters
}
\description{
Transform parameters to a meaningful and interpretable scale.
}
\author{
anonymous
}
