% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg.R
\name{plot.dd_curve}
\alias{plot.dd_curve}
\title{Plot Stock-Recruitment Curves for ADRM Objects}
\usage{
\method{plot}{dd_curve}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dd_curve}, usually the output of
\code{dd_curves()}.}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\description{
Automatically plots the stock-recruitment curves computed by
\code{dd_curves()}. Uses ggplot2 if available, otherwise base R graphics.
}
