% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg.R
\name{plot.marg_adrm}
\alias{plot.marg_adrm}
\title{Plot Marginal Relationships for ADRM Objects}
\usage{
\method{plot}{marg_adrm}(x, rug_data = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{marg_adrm}, usually the output of
\code{marg()}.}

\item{rug_data}{An optional \code{data.frame} containing the original data
to add a rug plot to the x-axis.}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\description{
Automatically plots the marginal relationship computed by
\code{marg()}. If \code{ggplot2} is installed, it returns a ggplot object;
otherwise, it falls back to base R graphics.
}
\details{
This default plotting method is restricted to outputs where exactly
one focal variable was evaluated (\code{length(variable) ==
  1}). Visualizing more complicated cases (e.g., 2D interactions) should be
handled manually by the user.
}
