% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{pop_dyn}
\alias{pop_dyn}
\title{Simulate population dynamics}
\usage{
pop_dyn(
  n_patches,
  n_time,
  n_ages,
  x_rec,
  ar_time,
  f_a_t,
  x_sv = NULL,
  m = 0.25,
  pars,
  init,
  init_type,
  movement = FALSE,
  adj_mat = NULL,
  mov_age = NULL
)
}
\arguments{
\item{n_patches}{number of patches}

\item{n_time}{number of timepoints}

\item{n_ages}{a \code{integer} indicating the number of (assumed)
age-classes.}

\item{x_rec}{matrix of environmental factors affecting recruitment}

\item{ar_time}{a \code{boolean} indicating whether an AR(1) term should be
included to the log-recruitment.}

\item{f_a_t}{fishing mortality}

\item{x_sv}{matrix of environmental factors affecting survival}

\item{m}{a \code{numeric} value indicating the natural mortality
instantaneous rate.}

\item{pars}{a named \code{list} of model parameters.}

\item{init}{initialization vector of length \code{n_ages - 1}}

\item{init_type}{type of initialization (integer between 0 and 5)}

\item{movement}{a \code{boolean} indicating whether movement should be
applied or not. If TRUE, than \code{pars} must have an element called
\code{zeta} indicating the probability of staying at a given patch between
two timepoints.}

\item{adj_mat}{a \code{n_patches} by \code{n_patches} row-standardized
adjacency \code{matrix}.}

\item{mov_age}{a \code{vector} of ages at which movement starts.}
}
\value{
an array of expected densities per age-group, patch, and timepoint.
}
\description{
Given a set of parameters, simulate population dynamics.
}
\author{
lcgodoy
}
