% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ppc}
\alias{ppc}
\alias{ppc.drmrmodels}
\title{Posterior Predictive Checks for DRM Models}
\usage{
ppc(x, ...)

\method{ppc}{drmrmodels}(
  x,
  type = c("density", "ecdf"),
  npost = 50,
  transform = c("none", "log1p"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{drmrmodels}.}

\item{...}{Additional graphical parameters passed to \code{plot}.}

\item{type}{Character string indicating the type of plot: \code{"density"}
(default) or \code{"ecdf"}.}

\item{npost}{Integer specifying the number of posterior draws to plot.
Default is 50. Usually, we do not use the total number of draws here
because the plot tends to get too heavy.}

\item{transform}{Character string specifying the scale for plotting. Either
\code{"none"} (default) or \code{"log1p"}.}
}
\description{
Generates posterior predictive check plots comparing observed
data to simulated data from the posterior predictive distribution.
}
