% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_drm}
\alias{predict_drm}
\title{Forecasts based on DRM.}
\usage{
predict_drm(drm, new_data, past_data, f_test, seed = 1, cores = 1)
}
\arguments{
\item{drm}{A \code{list} object containing the output from the \link{fit_drm}
function.}

\item{new_data}{a \code{data.frame} with the dataset at which we wish to
obtain predictions.}

\item{past_data}{a \code{data.frame} with the dataset last year used in
model fitting. Only needed when \code{f_test} is not missing or when
estimating survival.}

\item{f_test}{a \code{matrix} informing the instantaneous fishing mortality
rates at each age (columns) and timepoint (rows).}

\item{seed}{a seed used for the forecasts. Forecasts are obtained through
Monte Carlo samples from the posterior predictive distribution. Therefore,
a \code{seed} is needed to ensure the results' reproducibility.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{drm}, then four (or less) threads are recommended.}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for forecasting.
}
\description{
Considering a new dataset (across the same patches), computes
forecasts based on the DRM passed as \code{drm}.
}
\details{
The current version of the code assumes the data where forecasts
are needed is ordered by "patch" and "site" and, in addition, its patches
MUST be the same as the ones used to obtain the parameters' estimates from
the the \code{drm} object.
}
\author{
anonymous
}
