% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.adrm}
\alias{print.adrm}
\title{Print method for adrm objects}
\usage{
\method{print}{adrm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{adrm}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for adrm objects
}
