% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.drmrmodels}
\alias{print.drmrmodels}
\title{Print method for \code{adrm} and \code{sdm} objects}
\usage{
\method{print}{drmrmodels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{drmrmodels}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for \code{adrm} and \code{sdm} objects
}
