% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sdm}
\alias{print.sdm}
\title{Print method for sdm objects}
\usage{
\method{print}{sdm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sdm}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for sdm objects
}
