% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.adrm}
\alias{print.summary.adrm}
\title{Print method for summary.adrm}
\usage{
\method{print}{summary.adrm}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{summary.adrm}.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for summary.adrm
}
