% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.sdm}
\alias{print.summary.sdm}
\title{Print method for summary.adrm}
\usage{
\method{print}{summary.sdm}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{summary.sdm}.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for summary.adrm
}
