% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{prior_inits}
\alias{prior_inits}
\title{Generate initial values for MCMC from the prior}
\usage{
prior_inits(dat, chains, model = "drm")
}
\arguments{
\item{dat}{A list containing the data and prior parameters, typically
generated by \code{\link[=make_data]{make_data()}} or \code{\link[=make_data_sdm]{make_data_sdm()}}.}

\item{chains}{An integer specifying the number of MCMC chains to initialize.
Defaults to 4.}

\item{model}{A \code{character} string specifying the model type. Must be
either "drm" or "sdm". Defaults to "drm".}
}
\value{
A \code{list} of lists, where each inner list contains initial
values for one MCMC chain. The structure of each inner list is the same as
the output of \code{prior_sample()}.
}
\description{
Generates initial values for Markov Chain Monte Carlo (MCMC) by
sampling from the prior distributions of the model parameters. This helps
in starting the MCMC chains from different points in the parameter space.
}
\seealso{
\code{\link[=prior_sample]{prior_sample()}}
}
\author{
anonymous
}
