% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{prior_sample}
\alias{prior_sample}
\title{Generate samples from the prior distribution of model parameters}
\usage{
prior_sample(dat, model = "drm")
}
\arguments{
\item{dat}{A \code{list} containing the data and prior parameters, typically
generated by \code{\link[=make_data]{make_data()}} or \code{\link[=make_data_sdm]{make_data_sdm()}}.}

\item{model}{A \code{character} string specifying the model type. Must be
either "drm" or "sdm". Defaults to "drm".}
}
\value{
A \code{list} containing samples drawn from the prior distributions
of the model parameters.
}
\description{
Generates samples from the prior distributions of the model
parameters. This is primarily used for prior predictive checks or to
generate initial values for MCMC.
}
\seealso{
\code{\link[=prior_inits]{prior_inits()}}
}
\author{
anonymous
}
