% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rtt}
\alias{rtt}
\title{Random number generation from a truncated Student's t distribution}
\usage{
rtt(n, mean = 0, sd = 1, df = 3, range = c(0, Inf))
}
\arguments{
\item{n}{An \code{integer} specifying the number of random numbers to
generate.}

\item{mean}{A \code{numeric} scalar representing the mean of the underlying t
distribution. Defaults to 0.}

\item{sd}{A \code{numeric} scalar representing the standard deviation of the
underlying t distribution. Defaults to 1.}

\item{df}{A \code{numeric} scalar representing the degrees of freedom of the
underlying t distribution. Defaults to 3.}

\item{range}{A \code{numeric vector} of length 2 specifying the lower and
upper truncation bounds. Defaults to \code{c(0, Inf)}, indicating truncation
from below at 0.}
}
\value{
A \code{numeric vector} of length \code{n} containing random numbers
drawn from the specified truncated t distribution.
}
\description{
Generates random numbers from a truncated Student's t
distribution.
}
\details{
For details on the method used, see:
\url{https://stats.stackexchange.com/questions/567944/how-can-i-sample-from-a-shifted-and-scaled-student-t-distribution-with-a-specifi}
}
\seealso{
\code{\link[stats:rt]{stats::rt()}}, \code{\link[stats:qt]{stats::qt()}}
}
\author{
lcgodoy
}
