% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-run.R
\name{run_drm}
\alias{run_drm}
\title{Fit the dynamic range model.}
\usage{
run_drm(data_list, ...)
}
\arguments{
\item{data_list}{a \code{list} created using the \code{\link[=make_data]{make_data()}} function.}

\item{...}{Passed on to the \code{sample()} method of CmdStan model objects:
\url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\value{
A data frame of posterior summaries.
}
\description{
Fit the DRM Stan model and return posterior summary.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(sum_fl)
  make_data(y = sum_fl$y,
            time = sum_fl$year,
            site = sum_fl$patch,
            n_ages = 8,
            age_at_maturity = integer(0),
            m = 0.25) |>
  run_drm()
}
}
\seealso{
\code{\link[=make_data]{make_data()}}
}
\author{
lcgodoy
}
\concept{models}
