% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_modify}
\alias{safe_modify}
\title{Modifying a named list}
\usage{
safe_modify(original, replacements)
}
\arguments{
\item{original}{a named \code{list} with "original" parameters.}

\item{replacements}{a named \code{list} containing elements to be modified
in the \code{original} \code{list}.}
}
\value{
The updated \code{original} list.
}
\description{
Safely modifies a named \code{list}.
}
\details{
This function returns an error if any of the names found in the
\code{replacements} object are not in the \code{list} to be modified.
}
\author{
lcgodoy
}
