% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{sim_ar}
\alias{sim_ar}
\title{Simulate AR(1)}
\usage{
sim_ar(pars, n_time)
}
\arguments{
\item{pars}{a named \code{list} with two elements: \code{alpha}
(representing the temporal autocorrelation parameter) and \code{tau}
representing the conditional standard deviation of the AR(1) process.}

\item{n_time}{number of time points}
}
\value{
a \code{vector} of length \code{n_time} representing a realization
of a zero-mean AR(1) process.
}
\description{
Simulate AR(1)
}
\author{
lcgodoy
}
