% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{sim_dens}
\alias{sim_dens}
\title{Simulate response variable}
\usage{
sim_dens(x, pars)
}
\arguments{
\item{x}{an expected value for a simulated value.}

\item{pars}{a \code{list} of parameters.}
}
\value{
a \code{numeric} value representing a realization of the probability
distribution with mean \code{x} and parameters \code{pars}.
}
\description{
Given the expected value \code{x} and a list of parameters
(\code{pars}), this function simulates one realization of the probability
density function which is also specified by the list \code{pars}.
}
\author{
lcgodoy
}
