% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_adens}
\alias{summarise_adens}
\title{Extract and Summarize Age-Specific Densities}
\usage{
summarise_adens(lambda_obj, ages = NULL, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{lambda_obj}{A \code{CmdStanGQ} object returned by \code{lambda_drm}.}

\item{ages}{An optional vector of integers. If provided, the output is filtered
to include only these specific ages.}

\item{probs}{A numeric vector of quantiles to calculate in the summary.
Defaults to c(0.05, 0.5, 0.95).}
}
\value{
A \code{data.frame} containing the summary statistics and parsed indices
(age, time, patch).
}
\description{
Parses the output of \code{lambda_drm} to return a data frame of
age-specific densities, indexed by age, time, and patch.
}
