% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.adrm}
\alias{summary.adrm}
\title{Summary method for adrm objects}
\usage{
\method{summary}{adrm}(object, probs = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{object}{An object of class \code{adrm}.}

\item{probs}{A numeric vector of quantiles to calculate.
Defaults to c(0.05, 0.5, 0.95).}

\item{...}{Additional arguments passed to the cmdstanr summary method.}
}
\description{
Summary method for adrm objects
}
