% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.aesd}
\alias{summary.aesd}
\title{Summary method for \code{aesd} objects}
\usage{
\method{summary}{aesd}(object, probs = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{object}{An object of class \code{aesd}.}

\item{probs}{A numeric vector of quantiles to calculate.
Defaults to c(0.05, 0.5, 0.95).}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for \code{aesd} objects
}
