% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.CmdStanGQ}
\alias{summary.CmdStanGQ}
\title{Wrapper for the summary method for \code{CmdStanGQ} objects}
\usage{
\method{summary}{CmdStanGQ}(object, probs = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{object}{An object of class \code{CmdStanGQ}.}

\item{probs}{A numeric vector of quantiles to calculate.
Defaults to c(0.05, 0.5, 0.95).}

\item{...}{Additional arguments (ignored).}
}
\description{
Wrapper for the summary method for \code{CmdStanGQ} objects
}
