% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.drmrmodels}
\alias{summary.drmrmodels}
\title{Summary method for \code{adrm} and \code{sdm} objects}
\usage{
\method{summary}{drmrmodels}(object, probs = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{object}{An object of class \code{drmrmodels}.}

\item{probs}{A numeric vector of quantiles to calculate.
Defaults to c(0.05, 0.5, 0.95).}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for \code{adrm} and \code{sdm} objects
}
