% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.pred_drmr}
\alias{summary.pred_drmr}
\title{Summary method for \code{pred_drmr} objects}
\usage{
\method{summary}{pred_drmr}(object, probs = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{object}{An object of class \code{pred_drmr}.}

\item{probs}{A numeric vector of quantiles to calculate.
Defaults to c(0.05, 0.5, 0.95).}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for \code{pred_drmr} objects
}
