% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_aesd}
\alias{validate_aesd}
\title{Validate a aesd list object}
\usage{
validate_aesd(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
Nothing if validation is successful. Throws an error on failure.
}
\description{
Checks if an object is a valid \code{aesd} object by ensuring it is a
\code{list} and contains all required fields.
}
\keyword{internal}
