% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_drm_m}
\alias{validate_drm_m}
\title{Validate an adrm object}
\usage{
validate_drm_m(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
Nothing if validation is successful. Throws an error on failure.
}
\description{
Checks if an object is a valid \code{adrm} object by ensuring it is a \code{list}
and contains all required fields.
}
\keyword{internal}
