% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_pred_drmr}
\alias{validate_pred_drmr}
\title{Validate a pred_drmr list object}
\usage{
validate_pred_drmr(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
Nothing if validation is successful. Throws an error on failure.
}
\description{
Checks if an object is a valid \code{pred_drmr} object by ensuring it is a
\code{list} and contains all required fields.
}
\keyword{internal}
