% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_sdm}
\alias{validate_sdm}
\title{Validate an sdm object}
\usage{
validate_sdm(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
Nothing if validation is successful. Throws an error on failure.
}
\description{
Checks if an object is a valid \code{sdm} object by ensuring it is a \code{list}
and contains all required fields.
}
\keyword{internal}
