% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_priors}
\alias{default_priors}
\title{Default priors' hyperparameters}
\usage{
default_priors()
}
\value{
a \code{list} with (explain values)
}
\description{
Returns default priors' hyperparameters for the DRM model.
}
\author{
anonymous
}
