% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{default_priors}
\alias{default_priors}
\title{Default priors' hyperparameters}
\usage{
default_priors()
}
\value{
a `list` with (explain values)
}
\description{
Returns default priors' hyperparameters for the DRM model.
}
\details{
Based on Alexa's Summer Flounder project.
}
\author{
lcgodoy
}
