% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_toggles}
\alias{default_toggles}
\title{Default toggles}
\usage{
default_toggles()
}
\value{
a \code{list} with (explain values)
}
\description{
Returns default toggles for the DRM model.
}
\author{
anonymous
}
