% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{default_toggles}
\alias{default_toggles}
\title{Default toggles}
\usage{
default_toggles()
}
\value{
a `list` with (explain values)
}
\description{
Returns default toggles for the DRM model.
}
\details{
Based on Alexa's Summer Flounder project.
}
\author{
lcgodoy
}
