% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{drmr}
\alias{drmr-package}
\alias{drmr}
\title{drmr: Dynamic Range Models in Stan}
\description{
Fitting process-based species distribution models in stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pinskylab.github.io/drmr/}
  \item Report bugs at \url{https://github.com/pinskylab/drmr/issues/}
}

}
\author{
\strong{Maintainer}: Lucas da Cunha Godoy \email{lcgodoy@duck.com} (\href{https://orcid.org/0000-0003-4265-972X}{ORCID})

Authors:
\itemize{
  \item Lucas da Cunha Godoy \email{lcgodoy@duck.com} (\href{https://orcid.org/0000-0003-4265-972X}{ORCID})
  \item Malin Pinsky \email{mpinsky@ucsc.edu} (\href{https://orcid.org/0000-0002-8523-8952}{ORCID})
}

}
\concept{help}
