% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{drmr}
\alias{drmr-package}
\alias{drmr}
\title{drmr: Dynamic Range Models in Stan}
\description{
Fitting process-based species distribution models in stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/<removed-for-revie>}
  \item Report bugs at \url{https://github.com/<removed-for-review>}
}

}
\author{
\strong{Maintainer}: Author 1 \email{email@removed.for.review}

}
\concept{help}
