% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-run.R
\name{fit_drm}
\alias{fit_drm}
\title{Fit the dynamic range model.}
\usage{
fit_drm(
  .data,
  y_col,
  time_col,
  site_col,
  family = "gamma",
  formula_zero = ~1,
  formula_rec = ~1,
  formula_surv = NULL,
  seed,
  init = "cmdstan_default",
  algorithm = "nuts",
  algo_args = list(),
  ...
)
}
\arguments{
\item{.data}{A \code{data frame} containing the data for the model.}

\item{y_col}{A \code{character} specifying the name of the column in \code{.data}
that contains the response variable.}

\item{time_col}{A \code{character} specifying the name of the column in
\code{.data} that contains the time variable.}

\item{site_col}{A \code{character} specifying the name of the column in
\code{.data} that contains the site variable.}

\item{family}{a \code{character} specifying the family of the probability
distribution assumed for density. The options are: \itemize{ \item
\code{"gamma"} (default): gamma parametrized in terms of its mean; \item
\code{"lognormal"}: log-normal parametrized in terms of its mean; \item
\code{"loglogistic"}: log-logistic parametrized in terms of its median
(usual parametrization); \item \code{"lognormal_legacy"}: log-normal with
its usual parametrization; \item \code{"truncnorm"} a truncated normal
distribution. }}

\item{formula_zero}{A \code{formula} specifying the model for the zero
inflation component. Defaults to \code{~ 1} (intercept only).}

\item{formula_rec}{A \code{formula} specifying the model for the recruitment
component. Defaults to \code{~ 1} (intercept only).}

\item{formula_surv}{A \code{formula} specifying the model for the survival
component. If \code{NULL} (the default), no survival component is included.}

\item{seed}{An \code{integer} specifying the random number seed.}

\item{init}{A scalar specifying the initialization method. The default
("cmdstan_default") lets \code{cmdstan} initialize parameters.  Other
options include: a scalar greater than zero, say \code{x}, which
initializes all parameters uniformly between \code{-x} and \code{x};
\code{0}, which initializes all parameters at \code{0}; or "prior", which
initializes parameters by sampling from their priors.}

\item{algorithm}{a \code{character} specifying the algorithm used for
inference. Default is \code{nuts} (the default MCMC in Stan). The
remaining options are different flavors of variational bayes algorithms:
"vb" (for ADVI), "pathfinder" (for Pathfinder), "laplace" (normal
approximation centered at the mode of the posterior) or "optimize" for
(penalized) MLEs.}

\item{algo_args}{a \code{list} with arguments for the sampling
algorithms. For instance, \code{tol_rel_obj} for variational inference.
weight-at-age.}

\item{...}{Passed on to the \code{\link[=make_data]{make_data()}} function used to build the input
\code{list} for our \code{cmdstanr} model.}
}
\value{
An object of class \code{adrm} which is a \code{list} containing the
MCMC draws, the model data, the linear predictors formulas, and the
(response, time, site) column names.  Specifically: \itemize{ \item
\code{stanfit}: The MCMC draws from the fitted model.  \item \code{data}:
The data used to fit the model (as a list).  \item \code{formulas}: The
formulas used to create design matrices.  \item \code{cols}: Important
column names.  }
}
\description{
Fit the DRM Stan model
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(sum_fl)
  fit_drm(.data = sum_fl,
          y_col = "y",
          time_col = "year",
          site_col = "patch",
          seed = 2025)$stanfit$summary()
}
}
\seealso{
\code{\link[=make_data]{make_data()}}

Other models:
\code{\link[=fit_sdm]{fit_sdm()}}
}
\author{
lcgodoy
}
\concept{models}
