% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_gq.R
\name{fitted.sdm}
\alias{fitted.sdm}
\title{SDM fitted values}
\usage{
\method{fitted}{sdm}(object, cores = 1, ...)
}
\arguments{
\item{object}{A \code{sdm} object containing the output from the \link{fit_sdm}
function.}

\item{cores}{number of threads used for the forecast. If four chains were
used in the \code{sdm}, then four (or less) threads are recommended.}

\item{...}{additional parameters to be passed to \code{$generated_quantities}}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for the observed data.
}
\description{
Calculates the log-likelihood associated to a \code{sdm}
obkect.
}
\author{
lcgodoy
}
