% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes for ICAR spatial random effects}
\usage{
get_nodes(adj)
}
\arguments{
\item{adj}{adjacency \code{matrix}}
}
\value{
a \code{list}.
}
\description{
Get nodes for ICAR spatial random effects
}
\author{
lcgodoy
}
