% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{get_phi_hat}
\alias{get_phi_hat}
\title{Estimate phi}
\usage{
get_phi_hat(y, family)
}
\arguments{
\item{y}{a \code{numeric vector} of species' densities.}

\item{family}{a \code{character} specifying the family of the probability
distribution assumed for density. The options are: \itemize{ \item
\code{"gamma"} (default): gamma parametrized in terms of its mean; \item
\code{"lognormal"}: log-normal parametrized in terms of its mean; \item
\code{"loglogistic"}: log-logistic parametrized in terms of its mean.
\item \code{"lognormal_legacy"} (default): log-normal with its usual
parametrization; }}
}
\value{
a \code{numeric} scalar representing an estimate for phi
}
\description{
Estimate phi
}
\author{
lcgodoy
}
