% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{make_data}
\alias{make_data}
\title{Make data for DRM stan models}
\usage{
make_data(
  y,
  time,
  site,
  init_data = numeric(0),
  f_mort,
  m = -log(0.7),
  x_t,
  x_m,
  x_r,
  n_ages = 2,
  age_selectivity,
  ages_movement,
  adj_mat = matrix(0, ncol = 1, nrow = 1),
  mat = NULL,
  weight = NULL,
  .toggles,
  .priors,
  family = "gamma",
  reorder = TRUE,
  phi_hat = FALSE
)
}
\arguments{
\item{y}{a \code{numeric vector} of species' densities.}

\item{time}{a \code{vector} indicating the time point associated to each
element of \code{y}.}

\item{site}{a \code{vector} indicating the sites associated to each element
of \code{y}.}

\item{init_data}{an optional vector (of length n_ages - 1) to initialize the
population dynamics.}

\item{f_mort}{an optional \code{matrix} informing the instantaneous fishing
mortality rates at each age (columns) and timepoint (rows).}

\item{m}{a \code{numeric} value corresponding to the instantaneous natural
mortality rate. The default value for this is \code{-log(.7)}, as it
implies a survival rate of 0.70 between age classes.}

\item{x_t}{a design \code{matrix} of variables associated to the probability
of absence at each site/time.}

\item{x_m}{a design \code{matrix} of variables associated to survival.}

\item{x_r}{a design \code{matrix} of variables associated to recruitment.}

\item{n_ages}{an \code{integer} indicating the number of ages for the
underlying population dynamic model.}

\item{age_selectivity}{a \code{numeric vector} with \code{n_ages} elements,
where each element indicates the selectivity of a respective age. All the
elements of this vector must lie between 0 and 1.}

\item{ages_movement}{An \code{integer} or a \code{numeric vector} specifying
the ages at which individuals of the focal species are assumed to move. If
\code{ages_movement} is an integer, individuals younger than this age are
considered static (non-moving). If \code{ages_movement} is a numeric
vector of length \code{n_ages}, it indicates movement capability for each
age group. A value of \code{0} indicates the corresponding age group is
static, while \code{1} indicates movement is allowed. For example,
\code{c(0, 0, 1, 1, 0)} specifies that age groups 1, 2, and 5 are static,
while 3 and 4 are mobile.}

\item{adj_mat}{an adjacency \code{matrix} of dimensions \code{sites}
\eqn{\times} \code{sites}. Its elements are 1 if two sites are neighbors
and zero otherwise.}

\item{mat}{a \code{numeric vector} of length \code{n_ages} specifying the
maturity-at-age.}

\item{weight}{a \code{numeric vector} of length \code{n_ages} specifying the
weight-at-age.}

\item{.toggles}{a \code{list} of toggles for model components. The
components are: \itemize{ \item \code{rho_mu}: 1 to explicitly relate rho
to mu and 0 otherwise.  \item \code{cloglog}: 1 to use the complementary
log-log and 0 for the logit link function for the absence probabilities.
\item \code{movement}: 1 to allow for (adjacent) movement; 0 for static.
\item \code{est_surv}: 1 to estimate mortality and 0 otherwise.  \item
\code{est_init}: 1 to estimate initial values for lambda and 0 otherwise.
\item \code{minit}: 1 to use mortality to estimate initial age classes and
0 otherwise. \item \code{rec_dd}: 0 for Ricker, 1 for Beverton-Holt, and 2
(default) for no density dependence. \item \code{ar_re}: a \code{character}. It assumes one of
the following values: "none" no AR, "rec" AR(1) for recruitment, "surv"
AR(1) for survival (only works when \code{est_surv} is on), "dens" AR(1)
for density. \item \code{iid_re}: a \code{character}. It assumes one of
the following values: "none" no IID random effect, "rec" IID random effect
for recruitment, "surv" IID random effect for survival (only works when
\code{est_surv} is on), "dens" IID random effect for density.  \item
\code{sp_re}: a \code{character}. It assumes one of the following values:
"none" no ICAR random effect, "rec" ICAR random effect for recruitment,
"surv" ICAR random effect for survival (only works when \code{est_surv} is
on), "dens" ICAR random effect for density.}}

\item{.priors}{a \code{list} of priors hyperparameters.}

\item{family}{a \code{character} specifying the family of the probability
distribution assumed for density. The options are: \itemize{ \item
\code{"gamma"} (default): gamma parametrized in terms of its mean; \item
\code{"lognormal"}: log-normal parametrized in terms of its mean; \item
\code{"loglogistic"}: log-logistic parametrized in terms of its median
(usual parametrization); \item \code{"lognormal_legacy"}: log-normal with
its usual parametrization; }}

\item{reorder}{a \code{boolean} telling whether the data needs to be
reordered. The default is TRUE and means the data points will be ordered
by site and time, respectively.}

\item{phi_hat}{a \code{boolean} indicating whether the prior on \code{phi}
should be determined through the data (using \code{\link[=get_phi_hat]{get_phi_hat()}}).}
}
\value{
A \code{list} containing the data and settings to be used as input
for the Stan model.
}
\description{
This function creates the \code{list} used as the input for the \code{stan}
model.
}
\author{
lcgodoy
}
