% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{make_data}
\alias{make_data}
\title{Make data for DRM stan models}
\usage{
make_data(
  y,
  time,
  site,
  init_data = numeric(0),
  f_mort,
  m = -log(0.7),
  x_t,
  x_m,
  x_r,
  n_ages = 2,
  age_selectivity,
  ages_movement,
  adj_mat = matrix(0, ncol = 1, nrow = 1),
  .toggles,
  .priors,
  family = "gamma",
  reorder = TRUE,
  phi_hat = FALSE
)
}
\arguments{
\item{y}{a \code{numeric vector} of species' densities.}

\item{time}{an \code{vector} indicating the time point associated to each
element of \code{y}.}

\item{site}{an \code{vector} indicating the sites associated to each element
of \code{y}.}

\item{init_data}{an optional vector (of lengh n_ages - 1) to initialize the
population dynamics.}

\item{f_mort}{an optional \code{matrix} informing the instantaneous fishing
mortality rates at each age (columns) and timepoint (rows).}

\item{m}{a \code{numeric} value corresponding to the instantaneous natural
mortality rate. The default value for this is \code{-log(.7)}, as it
implies a survival rate of 0.70 between age classes.}

\item{x_t}{a design \code{matrix} of variables associated to the probability
of absence at each site/time.}

\item{x_m}{a design \code{matrix} of variables associated to survival.}

\item{x_r}{a design \code{matrix} of variables associated to recruitment.}

\item{n_ages}{an \code{integer} indicating the number of ages for the
underlying population dynamic model.}

\item{age_selectivity}{an \code{numeric vector} with \code{n_ages} elements,
where each element indicates the selectivity of a respective age. All the
elements of this vector must lie between 0 and 1.}

\item{ages_movement}{An \code{integer} or a \code{numeric vector} specifying
the ages at which individuals of the focal species are assumed to move. If
\code{ages_movement} is an integer, individuals younger than this age are
considered static (non-moving). If \code{ages_movement} is a numeric
vector of length \code{n_ages}, it indicates movement capability for each
age group. A value of \code{0} indicates the corresponding age group is
static, while \code{1} indicates movement is allowed. For example,
\code{c(0, 0, 1, 1, 0)} specifies that age groups 1, 2, and 5 are static,
while 3 and 4 are mobile.}

\item{adj_mat}{an adjacency \code{matrix} of dimensions \code{sites}
\eqn{\times} \code{sites}. Its elements are 1 if two sites are neighbors
and zero otherwise.}

\item{.toggles}{a \code{list} of toggles for model components. The
components are: \itemize{ \item \code{cloglog}: 1 to use the complementary
log-log and 0 for the logit link function for the absence probabilities.
\item \code{movement}: 1 to allow for (adjacent) moviment; 0 for static.
\item \code{est_surv}: 1 to estimate mortality and 0 otherwise.  \item
\code{est_init}: 1 to estimate initial values for lambda and 0 otherwise.
\item \code{time_ar}: 1 to incorporate an AR(1) process for recruitment.}}

\item{.priors}{a \code{list} of priors hyperparameters.}

\item{family}{a \code{character} specifying the family of the probability
distribution assumed for density. The options are: \itemize{ \item
\code{"gamma"} (default): gamma parametrized in terms of its mean; \item
\code{"lognormal"}: log-normal parametrized in terms of its mean; \item
\code{"loglogistic"}: log-logistic parametrized in terms of its mean.
\item \code{"lognormal_legacy"} (default): log-normal with its usual
parametrization; }}

\item{reorder}{a \code{boolean} telling whether the data needs to be
reordered. The default is TRUE and means the data points will be ordered
by site and time, respectively.}

\item{phi_hat}{a \code{boolean} indicating whether the prior on \code{phi}
should be determined through the data.}
}
\value{
a \code{list} to be used as the input for a \code{stan} model
}
\description{
This function creates the \code{list} used as the input for the \code{stan}
model.
}
\author{
lcgodoy
}
