% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marg}
\alias{marg}
\alias{marg_rec}
\alias{marg_pabs}
\alias{marg_surv}
\title{Relationships with covariates}
\usage{
marg_rec(drm, newdata)

marg_pabs(drm, newdata)

marg_surv(drm, newdata)
}
\arguments{
\item{drm}{the output of a \code{\link[=fit_drm]{fit_drm()}} call.}

\item{newdata}{a \code{data.frame} with the values of the environmental
variables at which we wish to estimate the recruitment, survival, or
absence probabilities.}
}
\value{
a \code{data.frame} with the samples the recruitment (or survival,
or absence probabilities) posterior distribution.
}
\description{
Relationships with covariates
}
\details{
the \code{marg_rec} function works to obtain relationships
regarding recruitment, while \code{marg_surv} and \code{marg_abs} evaluate
relationships with survival and absence probability, respectively.
}
\author{
lcgodoy
}
