% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gen.R
\name{model_sim}
\alias{model_sim}
\title{Generate a random sample from a model's predictive distribution given
a set of parameters}
\usage{
model_sim(dat, model, selectivity, pars, ...)
}
\arguments{
\item{dat}{A \code{list} containing the data and prior parameters, typically
generated by \code{\link[=make_data]{make_data()}} or \code{\link[=make_data_sdm]{make_data_sdm()}}.}

\item{model}{A \code{character} string specifying the model type. Must be
either "drm" or "sdm". Defaults to "drm".}

\item{selectivity}{a numeric \code{vector} with the same length as the
number of age-groups.}

\item{pars}{a named \code{list} of model parameters.}

\item{...}{parameters passed on to \code{pop_dyn}.}
}
\value{
A \code{list} containing samples drawn from the prior distributions
of the model parameters.
}
\description{
Generates samples from the predictive distribution of a model
given a set of parameters. This is primarily used for prior predictive
checks and simulation studies.
}
\details{
See
\href{https://forum.posit.co/t/how-to-solve-no-visible-binding-for-global-variable-note/28887/3}{this
link} for the \link[rlang:dot-data]{rlang::.data} import.
}
\seealso{
\code{\link[=prior_inits]{prior_inits()}}
}
\author{
anonymous
}
