% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{preddrm}
\alias{preddrm}
\alias{predict.adrm}
\alias{predict_drm}
\title{Predictions based on DRM.}
\usage{
\method{predict}{adrm}(
  object,
  new_data,
  past_data,
  f_test,
  type = "predictive",
  seed = 1,
  cores = 1,
  ...
)

predict_drm(...)
}
\arguments{
\item{object}{An \code{adrm} object containing the output from the
\code{\link[=fit_drm]{fit_drm()}} function.}

\item{new_data}{a \code{data.frame} with the dataset at which we wish to
obtain predictions.}

\item{past_data}{a \code{data.frame} with the dataset from the last year
used in model fitting. Only needed when \code{f_test} is not missing or
when estimating survival.}

\item{f_test}{a \code{matrix} informing the instantaneous fishing mortality
rates at each age (columns) and timepoint (rows).}

\item{type}{type of predictions to be computed. Admitted values are
\itemize{
\item \code{"predictive"} (default): posterior predictive distribution;
\item \code{"expected"}: theoretical mean of the posterior predictive distribution;
\item \code{"latent"}: latent density (i.e., disregarding the observation error);
}}

\item{seed}{a seed used for the predictions. predictions are obtained
through Monte Carlo samples from the posterior predictive
distribution. Therefore, a \code{seed} is needed to ensure the results'
reproducibility.}

\item{cores}{number of threads used to compute the predictions. If four
chains were used in the \code{drm}, then four (or less) threads are
recommended.}

\item{...}{params to be passed to \code{predict.adrm}}
}
\value{
an object of class \code{"CmdStanGQ"} containing samples for the
posterior predictive distribution for predictions.
}
\description{
Considering a new dataset (across the same patches), computes
predictions based on the DRM passed as \code{drm}.
}
\details{
The current version of the code assumes the data where predictions
are needed is ordered by "patch" and "site" and, in addition, its patches
MUST be the same as the ones used to obtain the parameters' estimates from
the the \code{drm} object.
}
\seealso{
\code{\link[=fit_drm]{fit_drm()}}
}
\author{
lcgodoy
}
