% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sdm}
\alias{predict.sdm}
\alias{predict_sdm}
\title{Predictions based on SDM.}
\usage{
\method{predict}{sdm}(object, new_data, type = "predictive", seed = 1, cores = 1, ...)

predict_sdm(...)
}
\arguments{
\item{object}{An \code{sdm} object containing the output of a \link{fit_sdm} call.}

\item{new_data}{a \code{data.frame} with the dataset at which we wish to
obtain predictions.}

\item{type}{type of predictions to be computed. Admitted values are
\itemize{
\item \code{"predictive"} (default): posterior predictive distribution;
\item \code{"expected"}: theoretical mean of the posterior predictive distribution;
\item \code{"latent"}: latent density (i.e., disregarding the observation error);
}}

\item{seed}{a seed used for the predictions. predictions are obtained
through Monte Carlo samples from the posterior predictive
distribution. Therefore, a \code{seed} is needed to ensure the results'
reproducibility.}

\item{cores}{number of threads used to compute the predictions. If four
chains were used in the \code{sdm}, then four (or less) threads are
recommended.}

\item{...}{params to be passed to \code{predict.sdm}}
}
\value{
An object of class \code{CmdStanGQ} (from the \code{instantiate}
package) containing samples for the posterior predictive distribution for
predictions.
}
\description{
Considering a new dataset (across the same sites), computes
predictions based on the SDM passed as \code{sdm}.
}
\details{
The current version of the code assumes the data where predictions
are needed is ordered by "site" and "site" and, in addition, its sites
MUST be the same as the ones used to obtain the parameters' estimates from
the the \code{sdm} object.
}
\seealso{
\code{\link[=fit_sdm]{fit_sdm()}}
}
\author{
lcgodoy
}
