% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.drmrmodels}
\alias{print.summary.drmrmodels}
\title{Print method for summary.spatial_model}
\usage{
\method{print}{summary.drmrmodels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.spatial_model}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for summary.spatial_model
}
