% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{rtn}
\alias{rtn}
\title{Random number generation from a truncated Normal distribution}
\usage{
rtn(n, mean = 0, sd = 1, range = c(0, Inf))
}
\arguments{
\item{n}{An \code{integer} specifying the number of random samples to
generate.}

\item{mean}{A \code{numeric} scalar representing the mean of the underlying
normal distribution. Defaults to 0.}

\item{sd}{A \code{numeric} scalar representing the standard deviation of the
underlying normal distribution. Defaults to 1.}

\item{range}{A \code{numeric vector} of length 2 specifying the lower and
upper truncation bounds. Defaults to \code{c(0, Inf)}, indicating
truncation from below at 0.}
}
\value{
A \code{numeric vector} of length \code{n} containing random numbers
drawn from the specified truncated normal distribution.
}
\description{
Generates random numbers from a truncated Normal distribution.
}
\seealso{
\code{\link[=rnorm]{rnorm()}}, \code{\link[=qnorm]{qnorm()}}
}
\author{
lcgodoy
}
