% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.adrm}
\alias{update.adrm}
\title{Update and Re-fit a DRM Model}
\usage{
\method{update}{adrm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{adrm}.}

\item{...}{Arguments to be updated in the new call (e.g., .data, formula_zero, iter_sampling).}
}
\value{
An updated \code{adrm} object.
}
\description{
Update and Re-fit a DRM Model
}
