% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.sdm}
\alias{update.sdm}
\title{Update and Re-fit a SDM Model}
\usage{
\method{update}{sdm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sdm}.}

\item{...}{Arguments to be updated in the new call.}
}
\value{
An updated \code{adrm} object or the unevaluated call.
}
\description{
Update and Re-fit a SDM Model
}
